backend/requirements.txt
fastapi>=0.115.8,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
slowapi>=0.1.9,<1.1.0
pandas>=2.2.3,<3.0.0
pinecone>=6.0.1,<7.0.0
requests>=2.32.3,<3.0.0
beautifulsoup4>=4.13.3,<5.0.0
tqdm>=4.67.1,<5.0.0
python-dotenv>=1.0.1,<2.0.0
boto3>=1.36.24,<2.0.0
openai>=1.63.2,<2.0.0